% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250113function_tcv.R
\name{multiDT}
\alias{multiDT}
\title{Thinning Cross-Validation for Determining the Number of Factors in Poisson Factor Models.}
\usage{
multiDT(x, K = 5, rmax = 8)
}
\arguments{
\item{x}{n by p matrix containing count responses, where n is the number of samples, p is the number of features.}

\item{K}{The number of folds for conducting thinning cross-validation. Default as 5.}

\item{rmax}{The maximum number of factors for conducting thinning cross-validation. Default as 8.}
}
\value{
The TCV error and TICV error. The number of factors corresponding to the smallest result is TCV estimator.
}
\description{
Thinning Cross-Validation for Determining the Number of Factors in Poisson Factor Models.
}
\author{
Zhijing Wang
}
